% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/there.R
\name{there}
\alias{there}
\title{Receive an inspirational message}
\usage{
there()
}
\value{
character
}
\description{
Randomly print a message to console, where the message is an inspirational saying.
}
\examples{
there()
}
