% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/there.R
\name{there}
\alias{there}
\title{Unquoted file referencing in project-oriented workflows}
\usage{
there(...)
}
\arguments{
\item{...}{unquoted folder or file names}
}
\value{
a character vector of length one -- a file path in the style of \href{https://here.r-lib.org/}{The here package}.
}
\description{
Unquoted file referencing in project-oriented workflows
}
\references{
\href{https://adv-r.hadley.nz/quasiquotation.html#quasi-motivation}{Motivation}
}
